[bug] 이미지 리사이징 실패
리사이징하고 좌표값 수정이 잘 안됨
시간이 더 들어갈거 같음 + 이미지 화질 저하 문제가 커서

이미지 자체 크기 키워서 넣는 걸로 대체할 예정

+++ 250304_WinAPI/UIMopHPBar.h
@@ -11,7 +11,7 @@ namespace UI
 		void Update() override;
 		void Render() override;
 
-		void SetPos(int dx, int dy) override;
+		void SetPos(float dx, float dy) override;
 		void SetMaxHP(int hp);
 		void SetHP(int hp);
 	protected:
